#feat import statment support&#fix memory leak

+++ include/buffer.h
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include <stdio.h>
 typedef void *Buffer;
-typedef Buffer (*Buffer_Alloc)(uint32_t size);
+typedef Buffer (*Buffer_Alloc)(size_t size);
 typedef void (*Buffer_Free)(Buffer buffer);
 void Buffer_init(Buffer_Alloc alloc, Buffer_Free free);
 Buffer Buffer_alloc(uint32_t size);
